# swagger/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: NodeJS Study Project
host: localhost:8080
basePath: /
...
paths:
...
  /api/products:
    x-swagger-router-controller: products
    get:
      operationId: getAll
      tags:
      - products
#      security:
#        - jwt_auth: []
      summary: Get all products
      description: Returns all products
      produces:
      - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/MessageError'
    post:
      operationId: add
      tags:
      - products
#      security:
#        - jwt_auth: []
      summary: Add new product
      description: Returns added product
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: body
        in: body
        description: Product
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/MessageError'
  /api/products/{id}:
    x-swagger-router-controller: products
    get:
      operationId: getOne
      tags:
      - products
#      security:
#        - jwt_auth: []
      summary: Get product by id
      description: Returns one product
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: ID of product
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/MessageError'
    delete:
      operationId: remove
      tags:
      - products
#      security:
#        - jwt_auth: []
      summary: Delete product by id
      description: Returns deleted product
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: ID of product
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/MessageDeleted'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/MessageError'
  /api/products/{id}/reviews:
    x-swagger-router-controller: products
    get:
      operationId: getAllReviews
      tags:
      - products
#      security:
#        - jwt_auth: []
      summary: Get product's reviews by it's id
      description: Returns all product's reviews
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: ID of product
        required: true
        type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: string
        404:
          description: Not found
          schema:
            $ref: '#/definitions/MessageError'
...
securityDefinitions:
  jwt_auth:
    type: apiKey
    in: header
    name: x-access-token
definitions:
...
  Product:
    type: object
    properties:
      name:
        type: string
      reviews:
        type: array
        items:
          type: string
      lastModifiedDate:
        type: string
        format: date-time
    required:
    - name
...
  MessageError:
    type: object
    properties:
      message:
        type: string
  MessageDeleted:
    type: object
    properties:
      ok:
        type: string
        enum:
        - deleted